# Placeholder interpolation