--- a/connection.go
+++ b/connection.go
@@ -34,21 +34,21 @@ type mysqlConn struct {
 }
 
 type config struct {
-	user                  string
-	passwd                string
-	net                   string
-	addr                  string
-	dbname                string
-	params                map[string]string
-	loc                   *time.Location
-	tls                   *tls.Config
-	timeout               time.Duration
-	collation             uint8
-	allowAllFiles         bool
-	allowOldPasswords     bool
-	clientFoundRows       bool
-	columnsWithAlias      bool
-	substitutePlaceholder bool
+	user              string
+	passwd            string
+	net               string
+	addr              string
+	dbname            string
+	params            map[string]string
+	loc               *time.Location
+	tls               *tls.Config
+	timeout           time.Duration
+	collation         uint8
+	allowAllFiles     bool
+	allowOldPasswords bool
+	clientFoundRows   bool
+	columnsWithAlias  bool
+	interpolateParams bool
 }
 
 // Handles parameters set in DSN after the connection is established
@@ -238,7 +238,7 @@ func (mc *mysqlConn) Exec(query string, args []driver.Value) (driver.Result, err
 		return nil, driver.ErrBadConn
 	}
 	if len(args) != 0 {
-		if !mc.cfg.substitutePlaceholder {
+		if !mc.cfg.interpolateParams {
 			return nil, driver.ErrSkip
 		}
 		// try client-side prepare to reduce roundtrip
@@ -289,7 +289,7 @@ func (mc *mysqlConn) Query(query string, args []driver.Value) (driver.Rows, erro
 		return nil, driver.ErrBadConn
 	}
 	if len(args) != 0 {
-		if !mc.cfg.substitutePlaceholder {
+		if !mc.cfg.interpolateParams {
 			return nil, driver.ErrSkip
 		}
 		// try client-side prepare to reduce roundtrip