@methane did awesome job in https://github.com/go-sql-driver/mysql/pull/297 and since I'm highly interested in this PR I wanted to test it. However https://github.com/go-sql-driver/mysql/pull/297 was out of sync with master ("This pull request contains merge conflicts that must be resolved.") so I merged changes from master.
Since I was already testing this. I've also:
* added some basic tests for escape functions (100% code coverage)
* exported escape functions (in case anyone would like to use them)
* added some basic sql injection tests (including sql_mode=NO_BACKSLASH_ESCAPES)
* referenced ported code to mysql source code (just in case if anyone like me;) would like to take a look if port is correct)

Feel free to merge this or throw away, I did this mostly to verify correctness.

One more time thanks for this @methane !