@arnehormann 2015-02-05
I'll hopefully get to review this soon.
Some remarks I already have from skimming it:

@arvenil add yourself to the AUTHORS file please.
@methane please close the other PR and reference this one if you are ok with arvenil taking over for you.

I will probably merge #311 first (it's pretty small), as this also hits `const.go`, you'll have to resync - sorry.

Can you change the tests so SELECT ... CAST is used instead of table writes? I don't know if that's applicable here, but I'd love to reduce the amount of disk IO during testing in the long run. See the DATE/TIME tests for example.

I don't like the Escape... functions, I prefer an externally controllable buffer and quotes embedded in the function.
Have a look at http://play.golang.org/p/ujxynUDVhE (extra points to you for a benchmark gist to compare speed & allocations).
I also think the EscapeQuoting tests are wrong. As a question: how would you be able to upload and retrieve a string `foo\x00bar` unchanged by the driver?