@methane 2015-02-06
I feel "interpolation" is better word than "substitution."  But I'm not sure since I'm native English.

@arvenil May I write optimized version of escaping function?