@arvenil 2015-02-06
@methane Yes, I've added you to https://github.com/arvenil/mysql/tree/placeholder as collaborator so you could push to this PR but do whatever suits you better.

I will address other questions and issues later, when time allows.