@arvenil 2015-02-07
Hi @arnehormann

> I also think the EscapeQuoting tests are wrong. As a question: how would you be able to upload and retrieve a string foo\x00bar unchanged by the driver?

For MySQL data is not changed, it is escaped. This is done exactly the same way how MySQL does it. Take a look at https://github.com/mysql/mysql-server/blob/mysql-5.7.5/mysys/charset.c#L823-L932 https://github.com/mysql/mysql-server/blob/mysql-5.7.5/mysys/charset.c#L963-L1038

To prove that data is not changed I've added `TestInsertRetrieveEscapedData`. You can see there that retrieved value equals inserted `foo \x00\n\r\x1a\"'\\`.