@methane 2015-02-09
I can reduce more allocates.

    * Split `escapeBackslash()` into `escapeBytesBackslash()` and `escapeStringBackslash()`. It reduces `string` to `[]byte` conversion.
    * Reserve 4 bytes in buffer for packet header.  Send packet with special function for it.

Both ideas makes mostly same pair of functions for `[]byte` and `string`.
I don't know I should do it or I should avoid duplicated code.