@methane 2015-02-10
Test failed on Go 1.1 since absence of `strings.IndexByte`.
Could we drop Go 1.1 support?