@methane 2015-02-10
I've tried extract duplicated code from escapeBytesBackslash and escapeStringBackslash.
Sadly, extracted function may not inlined and cause performance degrade.

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkInterpolation     2044          2448          +19.77%

benchmark                  old allocs     new allocs     delta
BenchmarkInterpolation     2              2              +0.00%

benchmark                  old bytes     new bytes     delta
BenchmarkInterpolation     448           448           +0.00%
```

[patch is here](https://gist.github.com/methane/8f2d085b2885c6efa7e6)