@arnehormann 2015-02-11
I'm fine with dropping Go 1.1, Gccgo is also at 1.2. @julienschmidt?