@julienschmidt 2015-02-11
Yep, it's fine. I'll also update [`parseDSN` in utitls.go](https://github.com/go-sql-driver/mysql/blob/8afc3be42072f568b4868b4afe97bff228a1fb32/utils.go#L82) then :wink: