@julienschmidt 2015-02-12
Let me also do a final review later before merging.