@julienschmidt 2015-02-12
@arvenil please add yourself to the AUTHORS file