@methane 2015-02-19
FYI, I've wrote blog post about optimizing this pull request.
http://methane.github.io/2015/02/reduce-allocation-in-go-code/